requirements.txt
django
django-allauth
pillow
reportlab
qrcode
pyzbar